+++ lib/widgets/pages/member_list/member_list_page.dart
@@ -4,7 +4,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:weforza/generated/l10n.dart';
 import 'package:weforza/model/debounce_search_delegate.dart';
 import 'package:weforza/model/member.dart';
-import 'package:weforza/widgets/pages/export_riders_page.dart';
+import 'package:weforza/widgets/pages/export_data_page/export_riders_page.dart';
 import 'package:weforza/widgets/pages/import_riders_page.dart';
 import 'package:weforza/widgets/pages/member_details/member_details_page.dart';
 import 'package:weforza/widgets/pages/member_form.dart';
